requirements.txt
boto3==1.18.31
botocore==1.21.31
click==8.0.1
docutils==0.16
fastjsonschema==2.15.1
Flask==2.0.1
flask-redis==0.4.0
Flask-SQLAlchemy==2.5.1
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
haversine==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.9.4
MarkupSafe==2.0.1
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.7.3
python-dotenv==0.19.0
redis==3.5.3
s3transfer==0.5.0
sacrud==0.3.7.post0
six==1.14.0
SQLAlchemy==1.4.23
sqlalchemy-mixins==1.5
transaction==3.0.1
typing==3.6.6
watchtower==0.5.5
Werkzeug== 2.0.0
zope.interface==5.4.0
uuid==1.30